Move `PDFSinglePageViewer` into the `web/pdf_viewer.js` file
With the previous commit, both of the `PDFViewer` and `PDFSinglePageViewer` clases are now small/simple enough that it no longer seems necessary to keep them in separate files.
